+++ NEWS.md
@@ -11,7 +11,8 @@
 
 ## Changes to functions
 
-* `link_inverse()` now supports `brmsfit` and `clm`-models.
+* `link_inverse()` now supports `brmsfit`, `multinom` and `clm`-models.
+* `p_value()` now supports `polr` and `multinom`-models.
 * `var_names()` now also cleans variable names from variables modelled with the `offset()` function.
 * `icc()`, `re_var()` and `get_re_var()` now support `brmsfit`-objects (models fitted with the *brms*-package).
 * For `fun = "weighted.mean"`, `typical_value()` now checks if vector of weights is of same length as `x`.